+++ LocalNuGetHelper/Public/PublishLocalPackage.ps1
@@ -1,4 +1,4 @@
-. $PSScriptRoot/../Private/PrepareLocalEnvironment.ps1
+﻿. $PSScriptRoot/../Private/PrepareLocalEnvironment.ps1
 . $PSScriptRoot/../Private/ExtractCsprojMetaData.ps1
 . $PSScriptRoot/../Private/CreatePackageToLocalFeed.ps1
 . $PSScriptRoot/../Private/UpdateGlobalPackageCache.ps1
@@ -6,7 +6,10 @@
 
 <#
     .SYNOPSIS
-    Packs your C# project of the current directory, publish it to your local NuGet feed and updates your global-package cache.
+    Packs your C# project, publish it to your local NuGet feed and updates your global-package cache.
+
+    .Parameter PackageProjectPath
+    The path to your package project (default = current location).
 
     .PARAMETER LocalFeedName
     The name for your local NuGet feed (default = Local NuGet feed).
@@ -23,24 +26,32 @@
     Procedure:
     1. Creates a directory for the local NuGet path under <LocalFeedPath>.
     2. Adds the directory as a NuGet source with the name <LocalFeedName>.
-    3. Packs the C# project of the current directory and puts the output into the local NuGet source.
+    3. Packs the C# project of the <PackageProjectDirectory> directory and puts the output into the local NuGet source.
     4. Purges local global-package cache of a maybe existing package version.
     5. Updates the global-package cache with the new package.
 
     .EXAMPLE
     Publish-LocalPackage
-    ...
-    ✓ Done
+
+    .EXAMPLE
+    Publish-LocalPackage -PackageProjectPath C:\dev\my-package-project
+
+    .EXAMPLE
+    Publish-LocalPackage -LocalFeedName "Fancy feed name" -LocalFeedPath C:\somedir\my-local-nuget-feed#
+
+    .EXAMPLE
+    Publish-LocalPackage -Watch
 
     .LINK
     PS Gallery page:  https://www.powershellgallery.com/packages/LocalNuGetHelper
 #>
 function Publish-LocalPackage {
-
-    # FIXME should variables be scoped to their respective function?
-
     [CmdletBinding()]
     param (
+        [Parameter()]
+        [String]
+        $PackageProjectPath = $(Get-Location),
+
         [Parameter()]
         [String]
         $LocalFeedName = "Local NuGet feed",
@@ -51,12 +62,16 @@ function Publish-LocalPackage {
 
         [Parameter()]
         [switch]
-        $watch = $false
+        $Watch = $false
     )
 
     $dotnetVersion = dotnet --version
     if (($null -eq $dotnetVersion) -or ($LASTEXITCODE -ne 0)) {
         Write-Error "No dotnet CLI available." -ErrorAction Stop
+    }
+
+    if ($PackageProjectPath -eq "") {
+        Write-Error "No package project path provided." -ErrorAction Stop
     }
 
     if ($LocalFeedName -eq "") {
@@ -67,8 +82,6 @@ function Publish-LocalPackage {
         Write-Error "No local feed path provided." -ErrorAction Stop
     }
 
-    # TODO maybe save a config and update the module description
-
     PrepareLocalEnvironment
 
     ($packageName, $packageVersion) = ExtractCsprojMetaData
@@ -77,7 +90,7 @@ function Publish-LocalPackage {
 
     UpdateGlobalPackageCache $packageName $packageVersion
 
-    if ($watch) {
+    if ($Watch) {
         RunPublishOnChanges
     }