+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ###  Added
 
 -   Automated integration test ("smoke test")
+-   `PacakgeProjectPath` parameter that will publish a package from a different directory than the current location
+-   `Watch` parameter that will publish a package continuously on each change
 
 ## [0.2.0] - 2021-07-25